+++ src/AutomationFoundation.NETCore.App/SigTermRunAsyncStrategy.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using AutomationFoundation.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace AutomationFoundation.Hosting
+namespace AutomationFoundation
 {
     /// <summary>
     /// Provides a run strategy which runs the host until a SIGTERM event has been received.
